Improve structure and make better messages in integration tests.
